+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/ref/RobotFileManager.java
@@ -4,15 +4,10 @@ import com.intellij.openapi.module.Module;
 import com.intellij.openapi.module.ModuleUtilCore;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.roots.ModuleRootManager;
-import com.intellij.openapi.util.SystemInfo;
 import com.intellij.openapi.vfs.VfsUtil;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiFile;
-import com.intellij.psi.PsiManager;
-import com.intellij.psi.search.FilenameIndex;
-import com.intellij.psi.search.GlobalSearchScope;
-import com.intellij.psi.search.ProjectScope;
 import com.intellij.psi.util.QualifiedName;
 import com.jetbrains.python.PyNames;
 import com.jetbrains.python.psi.PyFile;
@@ -21,14 +16,11 @@ import com.jetbrains.python.psi.stubs.PyModuleNameIndex;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.dto.VariableDto;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.DefinedVariable;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.util.ReservedVariable;
-import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Map;
-import java.util.Optional;
 
 public class RobotFileManager {
 
@@ -37,102 +29,15 @@ public class RobotFileManager {
 
     @Nullable
     public static VirtualFile findContentRootForFile(PsiFile file) {
-        VirtualFile sourceFile = file.getVirtualFile();
+        VirtualFile sourceFile = file.getViewProvider().getVirtualFile();
         Module moduleForFile = ModuleUtilCore.findModuleForFile(file);
         if (moduleForFile != null) {
-            ModuleRootManager moduleRootManager = ModuleRootManager.getInstance(moduleForFile);
-            VirtualFile[] contentRoots = moduleRootManager.getContentRoots();
-            VirtualFile relevantContentRoot = null;
+            VirtualFile[] contentRoots = ModuleRootManager.getInstance(moduleForFile).getContentRoots();
             for (VirtualFile contentRoot : contentRoots) {
                 if (VfsUtil.isAncestor(contentRoot, sourceFile, true)) {
-                    relevantContentRoot = contentRoot;
-                    break;
+                    return contentRoot;
                 }
             }
-            return relevantContentRoot;
-        }
-        return null;
-    }
-
-    @Nullable
-    public static PsiElement findElementInContext(@Nullable String elementName, @NotNull Project project, @NotNull PsiElement contextElement) {
-        if (elementName == null) {
-            return null;
-        }
-
-        if (!elementName.endsWith(".py")) {
-            return PythonResolver.resolveElement(elementName, project);
-        } else {
-            PsiFile psiFile = findPsiFile(elementName, project, ProjectScope.getContentScope(project), contextElement);
-            if (psiFile != null) {
-                return psiFile;
-            }
-            psiFile = findPsiFile(elementName, project, contextElement);
-            return psiFile;
-        }
-    }
-
-    @Nullable
-    private static PsiFile findPsiFile(@NotNull String filePath, @NotNull Project project, @NotNull PsiElement element) {
-        return findPsiFile(filePath, project, GlobalSearchScope.allScope(project), element);
-    }
-
-    @Nullable
-    private static PsiFile findPsiFile(@NotNull String filePath,
-                                       @NotNull Project project,
-                                       @NotNull GlobalSearchScope searchScope,
-                                       @NotNull PsiElement element) {
-        Path file = Path.of(filePath);
-        Collection<VirtualFile> foundFiles = FilenameIndex.getVirtualFilesByName(file.getFileName().toString(), searchScope);
-        if (!file.isAbsolute()) {
-            Optional<String> parentFilePathOpt = Optional.ofNullable(element.getContainingFile().getVirtualFile())
-                                                         .map(VirtualFile::getParent)
-                                                         .map(VirtualFile::getParent)
-                                                         .map(VirtualFile::getPath);
-            if (parentFilePathOpt.isEmpty()) {
-                return null;
-            }
-            String parentPath = parentFilePathOpt.get();
-            file = Path.of(parentPath, filePath);
-        }
-        String canonicalPath = file.toAbsolutePath().toString();
-        if (SystemInfo.isWindows) {
-            canonicalPath = canonicalPath.replace("\\", "/");
-        }
-        PsiManager psiManager = PsiManager.getInstance(project);
-        for (VirtualFile virtualFile : foundFiles) {
-            String virtualFilePath = virtualFile.getPath();
-            boolean samePath = !SystemInfo.isFileSystemCaseSensitive ? virtualFilePath.equalsIgnoreCase(canonicalPath) : virtualFilePath.equals(canonicalPath);
-            if (samePath) {
-                return psiManager.findFile(virtualFile);
-            }
-        }
-        return null;
-    }
-
-    @Nullable
-    public static PsiFile findPsiFiles(@NotNull String filePath, @Nullable VirtualFile sourceFile, @NotNull Project project) {
-        Path file = Path.of(filePath);
-        PsiManager psiManager = PsiManager.getInstance(project);
-        VirtualFile foundFile = null;
-        if (!file.isAbsolute()) {
-            if (sourceFile != null) {
-                foundFile = sourceFile.findFileByRelativePath("../" + filePath);
-            }
-        } else {
-            foundFile = VfsUtil.findFile(file, true);
-        }
-        if (foundFile == null) {
-            Collection<VirtualFile> foundFiles = FilenameIndex.getVirtualFilesByName(file.getFileName().toString(), GlobalSearchScope.allScope(project));
-            for (VirtualFile potentialFile : foundFiles) {
-                if (potentialFile.getPath().endsWith(filePath)) {
-                    foundFile = potentialFile;
-                    break;
-                }
-            }
-        }
-        if (foundFile != null) {
-            return psiManager.findFile(foundFile);
         }
         return null;
     }